+++ examples/two-chip.rs
@@ -222,13 +222,13 @@ impl<F: FieldExt> AddInstructions<F> for AddChip<F> {
                     || "lhs",
                     config.advice[0],
                     0,
-                    || a.value.ok_or(Error::SynthesisError),
+                    || a.value.ok_or(Error::Synthesis),
                 )?;
                 let rhs = region.assign_advice(
                     || "rhs",
                     config.advice[1],
                     0,
-                    || b.value.ok_or(Error::SynthesisError),
+                    || b.value.ok_or(Error::Synthesis),
                 )?;
                 region.constrain_equal(a.cell, lhs)?;
                 region.constrain_equal(b.cell, rhs)?;
@@ -239,7 +239,7 @@ impl<F: FieldExt> AddInstructions<F> for AddChip<F> {
                     || "lhs * rhs",
                     config.advice[0],
                     1,
-                    || value.ok_or(Error::SynthesisError),
+                    || value.ok_or(Error::Synthesis),
                 )?;
 
                 // Finally, we return a variable representing the output,
@@ -362,13 +362,13 @@ impl<F: FieldExt> MulInstructions<F> for MulChip<F> {
                     || "lhs",
                     config.advice[0],
                     0,
-                    || a.value.ok_or(Error::SynthesisError),
+                    || a.value.ok_or(Error::Synthesis),
                 )?;
                 let rhs = region.assign_advice(
                     || "rhs",
                     config.advice[1],
                     0,
-                    || b.value.ok_or(Error::SynthesisError),
+                    || b.value.ok_or(Error::Synthesis),
                 )?;
                 region.constrain_equal(a.cell, lhs)?;
                 region.constrain_equal(b.cell, rhs)?;
@@ -379,7 +379,7 @@ impl<F: FieldExt> MulInstructions<F> for MulChip<F> {
                     || "lhs * rhs",
                     config.advice[0],
                     1,
-                    || value.ok_or(Error::SynthesisError),
+                    || value.ok_or(Error::Synthesis),
                 )?;
 
                 // Finally, we return a variable representing the output,
@@ -457,7 +457,7 @@ impl<F: FieldExt> FieldInstructions<F> for FieldChip<F> {
                     || "private input",
                     config.advice[0],
                     0,
-                    || value.ok_or(Error::SynthesisError),
+                    || value.ok_or(Error::Synthesis),
                 )?;
                 num = Some(Number { cell, value });
                 Ok(())