+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to Rust's notion of
   - `Error` no longer implements `PartialEq`. Tests can check for specific error
     cases with `assert!(matches!(..))`.
   - `Error::IncompatibleParams` is now `Error::InvalidInstances`.
-  - `Error::TranscriptError` stores the underlying `io::Error`.
+  - `Error::OpeningError` is now `Error::Opening`.
+  - `Error::SynthesisError` is now `Error::Synthesis`.
+  - `Error::TranscriptError` is now `Error::Transcript`, and stores the
+    underlying `io::Error`.
 
 ### Removed
 - `halo2::arithmetic::BatchInvert` (use `ff::BatchInvert` instead).